super-webapp/requirements.txt
flask
opencv-python==4.1.2.30
pillow
tensorflow
numpy
scikit-image
matplotlib
keras